Repo patterns: allow spaces in a repo name
Repositories can have spaces in their names, but predefined patterns
don't allow this. This patch fixes it.

Also, `gitolite access` can check access to repositories with the list of them
on STDIN (used in src/triggers/post-compile/update-git-daemon-access-list;
it is an only usage in gitolite, but not in user scripts, yep). It expects
<reponame> [<username>] on a line. As far as <reponame> can contain spaces, it
gives 'DENIED' for such repos. So, make it expect <repo>[^I<username>] to avoid
error (^I is a TAB char).